refactor queryist utils
